If config file is not found or broken, launch with the default configuration, without creating any file
